add liq
