+++ dex/pair/src/pair_actions/mod.rs
@@ -1,3 +1,4 @@
+pub mod add_liq;
 pub mod common_methods;
 pub mod common_result_types;
 pub mod initial_liq;
