+++ dex/pair/tests/pair_setup/mod.rs
@@ -20,6 +20,7 @@ pub const USER_TOTAL_MEX_TOKENS: u64 = 5_000_000_000;
 pub const USER_TOTAL_WEGLD_TOKENS: u64 = 5_000_000_000;
 
 use pair::config::ConfigModule as PairConfigModule;
+use pair::pair_actions::add_liq::AddLiquidityModule;
 use pair::safe_price_view::*;
 use pair::*;
 use pausable::{PausableModule, State};
